feat: 단일 객체 리턴 타입 Optional로 변경

+++ backend/src/main/java/com/triply/barrierfreetrip/touristfacility/service/TouristFacilityHeartServiceImpl.java
@@ -8,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.util.Optional;
 
 
 @Service
@@ -18,17 +19,20 @@ public class TouristFacilityHeartServiceImpl implements TouristFacilityHeartServ
     private final TouristFacilityHeartRepository touristFacilityHeartRepository;
 
     public TouristFacilityHeart likes(Member member, String contentId, int likes) {
-        TouristFacility facility = facilityService.findByContentId(contentId);
+        Optional<TouristFacility> touristFacility = facilityService.findByContentId(contentId);
 
-        if (likes == 1) {  // 찜 추가
-            TouristFacilityHeart touristFacilityHeart = new TouristFacilityHeart(member, facility);
-            touristFacilityHeartRepository.save(touristFacilityHeart);
-            return touristFacilityHeart;
+        if (touristFacility.isPresent()) {
+            if (likes == 1) {  // 찜 추가
+                TouristFacilityHeart touristFacilityHeart = new TouristFacilityHeart(member, touristFacility.get());
+                touristFacilityHeartRepository.save(touristFacilityHeart);
+                return touristFacilityHeart;
 
-        } else {    // 찜 해제
-            TouristFacilityHeart prev = touristFacilityHeartRepository.findByIds(member, facility);
-            touristFacilityHeartRepository.delete(prev.getId());
+            } else {    // 찜 해제
+                TouristFacilityHeart prev = touristFacilityHeartRepository.findByIds(member, touristFacility.get());
+                touristFacilityHeartRepository.delete(prev.getId());
+            }
         }
+
         return null;
     }
 }
